+++ game/src/main/java/typeracer/game/observable/Observer.java
@@ -1,5 +1,7 @@
 package typeracer.game.observable;
 
+import typeracer.game.GameState;
+
 /** Interface for objects that receive updates from an Observable. */
 public interface Observer {
 
@@ -13,4 +15,27 @@ public interface Observer {
 
   /** Updates something. */
   void update();
+
+  /**
+   * Called when a new player joins.
+   *
+   * @param id of the new player
+   * @param newGameState the new game state
+   */
+  void updateNewPlayer(int id, GameState newGameState);
+
+  /**
+   * Called when the game state changes.
+   *
+   * @param newGameState the new game state
+   */
+  void updateState(GameState newGameState);
+
+  /**
+   * Called when a player is removed from the game.
+   *
+   * @param id of the removed player
+   * @param newGameState the new game state
+   */
+  void updateRemovedPlayer(int id, GameState newGameState);
 }